Implemento edición de servicios

+++ src/components/Cards/ServiceCard/ServiceCard.tsx
@@ -3,6 +3,9 @@ import Button from "../../../common/Button/Button"
 import { useFetchData } from "../../../hooks/useFetchData"
 import { BACKEND_API_URL } from "../../../config"
 import { notifyError, notifySuccess } from "../../../utils/notifications"
+import { confirmDelete } from "../../../utils/alerts"
+import ModalForm from "../../ModalForm/ModalForm"
+import { useState } from "react"
 
 interface Props {
     id: string
@@ -11,29 +14,56 @@ interface Props {
     title: string
     description: string
     onDeleteService: (id: string, scheduledAppointmentsToDelete: string[]) => void
+    onUpdateService: (data: { [key: string]: any }) => void
 }
 
-const ServiceCard: React.FC<Props> = ({ id, duration, price, title, description, onDeleteService }) => {
+const ServiceCard: React.FC<Props> = ({ id, duration, price, title, description, onDeleteService, onUpdateService }) => {
 
     const { isLoading, error, fetchData } = useFetchData(
         `${BACKEND_API_URL}/services/delete-service/${id}`,
-        "POST",
+        "DELETE",
         true
     )
 
-    if (error) {
-        console.error(error)
+    const { isLoading: isLoadingUpdate, error: errorUpdate, fetchData: fetchDataUpdate } = useFetchData(
+        `${BACKEND_API_URL}/services/edit-service/${id}`,
+        "PUT",
+        true
+    )
+
+    const [isModalOpen, setIsModalOpen] = useState<boolean>(false)
+
+    if (error || errorUpdate) {
+        console.error(error || errorUpdate)
         notifyError('Error del servidor: Inténtalo de nuevo más tarde')
     }
 
     const deleteService = async () => {
-        const response = await fetchData({})
-        console.log(response)
-        if (response?.data) {
-            onDeleteService(id, response.appointmentsToDelete)
-            notifySuccess("Servicio eliminado")
+        const deleteConfirmed = await confirmDelete({
+            question: "¿Seguro que desea eliminar el servicio?",
+            mesasge: "Al eliminar el servicio se eliminarán todos los turnos pendientes de clientes relacionados a este servicio.",
+            icon: "warning",
+            confirmButtonText: "Eliminar servicio",
+            cancelButtonText: "Cancelar"
+        })
+        if (deleteConfirmed) {
+            const response = await fetchData({})
+            if (response?.data) {
+                onDeleteService(id, response.appointmentsToDelete)
+                notifySuccess("Servicio eliminado")
+            }
+            if (response?.error) notifyError("Error al eliminar el servicio")
         }
-        if (response?.error) notifyError("Error al eliminar el servicio")
+    }
+
+    const updateService = async (data: { [key: string]: any }) => {
+        const response = await fetchDataUpdate(data)
+        setIsModalOpen(false)
+        if (response?.data) {
+            onUpdateService(response.data)
+            notifySuccess("Servicio actualizado")
+        }
+        if (response?.error) notifyError("Error al actualizar el servicio")
     }
 
     return (
@@ -48,14 +78,15 @@ const ServiceCard: React.FC<Props> = ({ id, duration, price, title, description,
                 <Button
                     fontSize="1.2rem"
                     padding=".8rem"
-                    disabled={isLoading}
+                    disabled={isLoading || isLoadingUpdate}
                 >
                     Habilitar turnos
                 </Button>
                 <Button
                     fontSize="1.2rem"
                     padding=".8rem"
-                    disabled={isLoading}
+                    disabled={isLoading || isLoadingUpdate}
+                    onSubmit={() => setIsModalOpen(true)}
                 >
                     Editar servicio
                 </Button>
@@ -63,11 +94,24 @@ const ServiceCard: React.FC<Props> = ({ id, duration, price, title, description,
                     fontSize="1.2rem"
                     padding=".8rem"
                     onSubmit={deleteService}
-                    disabled={isLoading}
+                    disabled={isLoading || isLoadingUpdate}
                 >
                     Eliminar servicio
                 </Button>
             </div>
+            <ModalForm
+                title="Editar servicio"
+                isOpen={isModalOpen}
+                inputs={[
+                    { type: "text", name: "title", placeholder: "Título", value: title },
+                    { type: "text", name: "description", placeholder: "Descripción", value: description },
+                    { type: "number", name: "price", placeholder: "Precio", value: price },
+                    { type: "number", name: "duration", placeholder: "Duración", value: duration }
+                ]}
+                onClose={() => setIsModalOpen(false)}
+                onSubmitForm={(data) => updateService(data)}
+                disabledButtons={isLoading}
+            />
         </div>
     );
 }
